3Dview settings are back
